engine: extract esm/{view_mover,tours/tour_player}.js
