+++ engine/esm/transpiled.js
@@ -32,11 +32,13 @@ import {
   globalScriptInterface,
   globalWWTControl,
   makeNewHipsProperties,
+  setManagerVisibleLayerList,
   set_freestandingMode,
   set_globalScriptInterface,
   set_globalWWTControl,
   set_makeNewFolder,
   set_makeNewHipsProperties,
+  set_setManagerVisibleLayerList,
 } from "./data_globals.js";
 
 import { BlendState } from "./blend_state.js";
@@ -121,6 +123,7 @@ import { MinorPlanets } from "./minor_planets.js";
 import { TileCache } from "./tile_cache.js";
 import { Tour } from "./tour.js";
 import { VideoOutputType } from "./video_output_type.js";
+import { ViewMoverSlew, ViewMoverKenBurnsStyle } from "./view_mover.js";
 import { ColorPicker } from "./utilities/color_picker.js";
 import { Dialog } from "./utilities/dialog.js";
 import { Histogram } from "./utilities/histogram.js";
@@ -137,10 +140,11 @@ import { Table } from "./layers/table.js";
 import { VoTable } from "./layers/vo_table.js";
 
 import { FileCabinet } from "./tours/file_cabinet.js";
-import { Overlay, TextOverlay } from "./tours/overlay.js";
+import { Overlay } from "./tours/overlay.js";
 import { TourStop, LayerInfo, UndoTourStopChange } from "./tours/tour_stop.js";
 import { Undo, UndoTourSlidelistChange } from "./tours/undo.js";
 import { TourEditor, TourEdit, SoundEditor, TourStopList, OverlayList, TimeLine } from "./tours/tour_editor.js";
+import { TourPlayer, MasterTime } from "./tours/tour_player.js";
 
 
 // wwtlib.PointType
@@ -1491,215 +1495,6 @@ var FolderBrowser$ = {
 
 registerType("FolderBrowser", [FolderBrowser, FolderBrowser$, null]);
 
-// wwtlib.ViewMoverKenBurnsStyle
-//
-// This was defined in `IViewMover.cs`, which we've folded into `interfaces.js`.
-
-export function ViewMoverKenBurnsStyle(from, to, time, fromDateTime, toDateTime, type) {
-  this.interpolationType = 0;
-  this.fastDirectionMove = false;
-  this._toTargetTime = 0;
-  this._dateTimeSpan = 0;
-  this._complete = false;
-  this._midpointFired = false;
-  this.interpolationType = type;
-  if (Math.abs(from.lng - to.lng) > 180) {
-    if (from.lng > to.lng) {
-      from.lng -= 360;
-    } else {
-      from.lng += 360;
-    }
-  }
-  this._fromDateTime = fromDateTime;
-  this._toDateTime = toDateTime;
-  this._dateTimeSpan = toDateTime - fromDateTime;
-  this._from = from.copy();
-  this._to = to.copy();
-  this._fromTime = SpaceTimeController.get_metaNow();
-  this._toTargetTime = time;
-}
-
-var ViewMoverKenBurnsStyle$ = {
-  get_complete: function () {
-    return this._complete;
-  },
-
-  get_currentPosition: function () {
-    var elapsed = SpaceTimeController.get_metaNow() - this._fromTime;
-    var elapsedSeconds = (elapsed) / 1000;
-    var alpha = elapsedSeconds / this._toTargetTime;
-    if (!this._midpointFired && alpha >= 0.5) {
-      this._midpointFired = true;
-      if (this._midpoint != null) {
-        this._midpoint();
-      }
-    }
-    if (alpha >= 1) {
-      alpha = 1;
-      this._complete = true;
-      return this._to.copy();
-    }
-    if (Settings.get_active().get_galacticMode() && globalRenderContext.space) {
-      return CameraParameters.interpolateGreatCircle(this._from, this._to, alpha, this.interpolationType, this.fastDirectionMove);
-    }
-    return CameraParameters.interpolate(this._from, this._to, alpha, this.interpolationType, this.fastDirectionMove);
-  },
-
-  get_currentDateTime: function () {
-    var elapsed = SpaceTimeController.get_metaNow() - this._fromTime;
-    var elapsedSeconds = (elapsed) / 1000;
-    var alpha = elapsedSeconds / this._toTargetTime;
-    var delta = this._dateTimeSpan * alpha;
-    var retDate = new Date(this._fromDateTime.getTime() + ss.truncate(delta));
-    return retDate;
-  },
-
-  get_midpoint: function () {
-    return this._midpoint;
-  },
-
-  set_midpoint: function (value) {
-    this._midpoint = value;
-    return value;
-  },
-
-  get_moveTime: function () {
-    return this._toTargetTime;
-  }
-};
-
-registerType("ViewMoverKenBurnsStyle", [ViewMoverKenBurnsStyle, ViewMoverKenBurnsStyle$, null, IViewMover]);
-
-// wwtlib.ViewMoverSlew
-//
-// This was defined in `IViewMover.cs`, which we've folded into `interfaces.js`.
-
-export function ViewMoverSlew() {
-  this._upTargetTime = 0;
-  this._downTargetTime = 0;
-  this._toTargetTime = 0;
-  this._upTimeFactor = 0.6;
-  this._downTimeFactor = 0.6;
-  this._travelTimeFactor = 7;
-  this._midpointFired = false;
-  this._complete = false;
-}
-
-ViewMoverSlew.create = function (from, to) {
-  var temp = new ViewMoverSlew();
-  temp.init(from, to);
-  return temp;
-};
-
-ViewMoverSlew.createUpDown = function (from, to, upDowFactor) {
-  var temp = new ViewMoverSlew();
-  temp._upTimeFactor = temp._downTimeFactor = upDowFactor;
-  temp.init(from.copy(), to.copy());
-  return temp;
-};
-
-var ViewMoverSlew$ = {
-  init: function (from, to) {
-    if (Math.abs(from.lng - to.lng) > 180) {
-      if (from.lng > to.lng) {
-        from.lng -= 360;
-      }
-      else {
-        from.lng += 360;
-      }
-    }
-    if (to.zoom <= 0) {
-      to.zoom = 360;
-    }
-    if (from.zoom <= 0) {
-      from.zoom = 360;
-    }
-    this._from = from;
-    this._to = to;
-    this._fromTime = SpaceTimeController.get_metaNow();
-    var zoomUpTarget = 360;
-    var travelTime;
-    var lngDist = Math.abs(from.lng - to.lng);
-    var latDist = Math.abs(from.lat - to.lat);
-    var distance = Math.sqrt(latDist * latDist + lngDist * lngDist);
-    zoomUpTarget = (distance / 3) * 20;
-    if (zoomUpTarget > 360) {
-      zoomUpTarget = 360;
-    }
-    if (zoomUpTarget < from.zoom) {
-      zoomUpTarget = from.zoom;
-    }
-    travelTime = (distance / 180) * (360 / zoomUpTarget) * this._travelTimeFactor;
-    var rotateTime = Math.max(Math.abs(from.angle - to.angle), Math.abs(from.rotation - to.rotation));
-    var logDistUp = Math.max(Math.abs(Util.logN(zoomUpTarget, 2) - Util.logN(from.zoom, 2)), rotateTime);
-    this._upTargetTime = this._upTimeFactor * logDistUp;
-    this._downTargetTime = this._upTargetTime + travelTime;
-    var logDistDown = Math.abs(Util.logN(zoomUpTarget, 2) - Util.logN(to.zoom, 2));
-    this._toTargetTime = this._downTargetTime + Math.max((this._downTimeFactor * logDistDown), rotateTime);
-    this._fromTop = from.copy();
-    this._fromTop.zoom = zoomUpTarget;
-    this._fromTop.angle = (from.angle + to.angle) / 2;
-    this._fromTop.rotation = (from.rotation + to.rotation) / 2;
-    this._toTop = to.copy();
-    this._toTop.zoom = this._fromTop.zoom;
-    this._toTop.angle = this._fromTop.angle;
-    this._toTop.rotation = this._fromTop.rotation;
-  },
-
-  get_complete: function () {
-    return this._complete;
-  },
-
-  get_currentPosition: function () {
-    var elapsed = SpaceTimeController.get_metaNow() - this._fromTime;
-    var elapsedSeconds = (elapsed) / 1000;
-    if (elapsedSeconds < this._upTargetTime) {
-      return CameraParameters.interpolate(this._from, this._fromTop, elapsedSeconds / this._upTargetTime, 3, false);
-    } else if (elapsedSeconds < this._downTargetTime) {
-      elapsedSeconds -= this._upTargetTime;
-      if (Settings.get_active().get_galacticMode() && globalRenderContext.space) {
-        return CameraParameters.interpolateGreatCircle(this._fromTop, this._toTop, elapsedSeconds / (this._downTargetTime - this._upTargetTime), 3, false);
-      }
-      return CameraParameters.interpolate(this._fromTop, this._toTop, elapsedSeconds / (this._downTargetTime - this._upTargetTime), 3, false);
-    } else {
-      if (!this._midpointFired) {
-        this._midpointFired = true;
-        if (this._midpoint != null) {
-          this._midpoint();
-        }
-      }
-      elapsedSeconds -= this._downTargetTime;
-      var alpha = elapsedSeconds / (this._toTargetTime - this._downTargetTime);
-      if (alpha > 1) {
-        alpha = 1;
-        this._complete = true;
-        return this._to.copy();
-      }
-      return CameraParameters.interpolate(this._toTop, this._to, alpha, 3, false);
-    }
-  },
-
-  get_currentDateTime: function () {
-    SpaceTimeController.updateClock();
-    return SpaceTimeController.get_now();
-  },
-
-  get_midpoint: function () {
-    return this._midpoint;
-  },
-
-  set_midpoint: function (value) {
-    this._midpoint = value;
-    return value;
-  },
-
-  get_moveTime: function () {
-    return this._toTargetTime;
-  }
-};
-
-registerType("ViewMoverSlew", [ViewMoverSlew, ViewMoverSlew$, null, IViewMover]);
-
 
 // The `Layer.fromXml` function, which needs to be aware of a bunch of layer
 // subclasses.
@@ -2489,6 +2284,8 @@ LayerManager.setVisibleLayerList = function (list) {
     }
   }
 };
+
+set_setManagerVisibleLayerList(LayerManager.setVisibleLayerList);
 
 LayerManager._preDraw = function (renderContext, opacity, astronomical, referenceFrame, nested) {
   if (!ss.keyExists(LayerManager.get_allMaps(), referenceFrame)) {
@@ -6318,7 +6115,7 @@ var TourEditTab$ = {
       this.tourStopList.selectedItem = this._tour.get_currentTourstopIndex();
       this.musicTrack.target = this._tour.get_currentTourStop();
       this.voiceTrack.target = this._tour.get_currentTourStop();
-      LayerManager.setVisibleLayerList(this._tour.get_currentTourStop().layers);
+      setManagerVisibleLayerList(this._tour.get_currentTourStop().layers);
     }
     this.setEditMode(this._tour.get_editMode());
     return value;
@@ -6398,7 +6195,7 @@ var TourEditTab$ = {
       globalWWTControl.gotoTarget(ts.get_target(), false, true, false);
       this._tour.get_currentTourStop().set_tweenPosition(0);
       this._tour.get_currentTourStop()._updateLayerOpacity();
-      LayerManager.setVisibleLayerList(this._tour.get_currentTourStop().layers);
+      setManagerVisibleLayerList(this._tour.get_currentTourStop().layers);
     }
   },
 
@@ -6688,7 +6485,7 @@ var TourEditTab$ = {
       SpaceTimeController.set_now(this._tour.get_currentTourStop().get_startTime());
       SpaceTimeController.set_syncToClock(false);
       this._tour.get_currentTourStop().set_tweenPosition(0);
-      LayerManager.setVisibleLayerList(this._tour.get_currentTourStop().layers);
+      setManagerVisibleLayerList(this._tour.get_currentTourStop().layers);
       this.tourStopList.refresh();
     }
   },
@@ -6701,7 +6498,7 @@ var TourEditTab$ = {
       globalRenderContext.set_solarSystemTrack(this._tour.get_currentTourStop().get_endTarget().get_target());
       SpaceTimeController.set_now(this._tour.get_currentTourStop().get_endTime());
       this._tour.get_currentTourStop().syncSettings();
-      LayerManager.setVisibleLayerList(this._tour.get_currentTourStop().layers);
+      setManagerVisibleLayerList(this._tour.get_currentTourStop().layers);
       SpaceTimeController.set_syncToClock(false);
       this.tourStopList.refresh();
       this.tourEditorUI.clearSelection();
@@ -7112,610 +6909,6 @@ var TourEditTab$ = {
 };
 
 registerType("TourEditTab", [TourEditTab, TourEditTab$, null]);
-
-// wwtlib.TourPlayer
-
-export function TourPlayer() {
-  this._overlayBlend = BlendState.create(false, 1000);
-  this._tour = null;
-  this._onTarget = false;
-  this._currentMasterSlide = null;
-  this._callStack = new ss.Stack();
-  this._leaveSettingsWhenStopped = false;
-}
-
-TourPlayer._playing = false;
-TourPlayer._switchedToFullScreen = false;
-TourPlayer.noRestoreUIOnStop = false;
-
-TourPlayer.get_playing = function () {
-  return TourPlayer._playing;
-};
-
-TourPlayer.set_playing = function (value) {
-  TourPlayer._playing = value;
-  return value;
-};
-
-TourPlayer.add_tourEnded = function (value) {
-  TourPlayer.__tourEnded = ss.bindAdd(TourPlayer.__tourEnded, value);
-};
-
-TourPlayer.remove_tourEnded = function (value) {
-  TourPlayer.__tourEnded = ss.bindSub(TourPlayer.__tourEnded, value);
-};
-
-var TourPlayer$ = {
-  render: function (renderContext) {
-    if (this._tour == null || this._tour.get_currentTourStop() == null || !TourPlayer._playing) {
-      return;
-    }
-    renderContext.save();
-    this.updateSlideStates();
-    if (!this._onTarget) {
-      this._slideStartTime = ss.now();
-      if (renderContext.onTarget(this.get_tour().get_currentTourStop().get_target())) {
-        this._onTarget = true;
-        this._overlayBlend.set_state(!this.get_tour().get_currentTourStop().get_fadeInOverlays());
-        this._overlayBlend.set_targetState(true);
-        if (this._tour.get_currentTourStop().get_musicTrack() != null) {
-          this._tour.get_currentTourStop().get_musicTrack().seek(0);
-          this._tour.get_currentTourStop().get_musicTrack().play();
-        }
-        if (this._tour.get_currentTourStop().get_voiceTrack() != null) {
-          this._tour.get_currentTourStop().get_voiceTrack().seek(0);
-          this._tour.get_currentTourStop().get_voiceTrack().play();
-        }
-        var caption = '';
-        var $enum1 = ss.enumerate(this._tour.get_currentTourStop().get_overlays());
-        while ($enum1.moveNext()) {
-          var overlay = $enum1.current;
-          if (overlay.get_name().toLowerCase() === 'caption') {
-            var text = ss.safeCast(overlay, TextOverlay);
-            if (text != null) {
-              caption = text.textObject.text;
-            }
-          }
-          overlay.play();
-        }
-        LayerManager.setVisibleLayerList(this._tour.get_currentTourStop().layers);
-        if (this._tour.get_currentTourStop().get_endTarget() != null && this._tour.get_currentTourStop().get_endTarget().get_zoomLevel() !== -1) {
-          if (this._tour.get_currentTourStop().get_target().get_type() === 4) {
-          }
-          renderContext.viewMover = new ViewMoverKenBurnsStyle(this._tour.get_currentTourStop().get_target().get_camParams(), this._tour.get_currentTourStop().get_endTarget().get_camParams(), this._tour.get_currentTourStop().get_duration() / 1000, this._tour.get_currentTourStop().get_startTime(), this._tour.get_currentTourStop().get_endTime(), this._tour.get_currentTourStop().get_interpolationType());
-        }
-        Settings.tourSettings = this._tour.get_currentTourStop();
-        SpaceTimeController.set_now(this._tour.get_currentTourStop().get_startTime());
-        SpaceTimeController.set_syncToClock(false);
-        globalScriptInterface._fireSlideChanged(caption);
-      }
-    }
-    if (renderContext.gl != null) {
-      renderContext.setupMatricesOverlays();
-      if (this._currentMasterSlide != null) {
-        var $enum2 = ss.enumerate(this._currentMasterSlide.get_overlays());
-        while ($enum2.moveNext()) {
-          var overlay = $enum2.current;
-          overlay.set_tweenFactor(1);
-          overlay.draw3D(renderContext, false);
-        }
-      }
-      if (this._onTarget) {
-        var $enum3 = ss.enumerate(this._tour.get_currentTourStop().get_overlays());
-        while ($enum3.moveNext()) {
-          var overlay = $enum3.current;
-          if (overlay.get_name().toLowerCase() !== 'caption' || globalScriptInterface.get_showCaptions()) {
-            overlay.set_tweenFactor(CameraParameters.easeCurve(this._tour.get_currentTourStop().get_tweenPosition(), (overlay.get_interpolationType() === 5) ? this._tour.get_currentTourStop().get_interpolationType() : overlay.get_interpolationType()));
-            overlay.draw3D(renderContext, false);
-          }
-        }
-      }
-      renderContext.restore();
-    } else {
-      renderContext.device.scale(renderContext.height / 1116, renderContext.height / 1116);
-      var aspectOrig = 1920 / 1116;
-      var aspectNow = renderContext.width / renderContext.height;
-      renderContext.device.translate(-((1920 - (aspectNow * 1116)) / 2), 0);
-      if (this._currentMasterSlide != null) {
-        var $enum4 = ss.enumerate(this._currentMasterSlide.get_overlays());
-        while ($enum4.moveNext()) {
-          var overlay = $enum4.current;
-          overlay.set_tweenFactor(1);
-          overlay.draw3D(renderContext, false);
-        }
-      }
-      if (this._onTarget) {
-        var $enum5 = ss.enumerate(this._tour.get_currentTourStop().get_overlays());
-        while ($enum5.moveNext()) {
-          var overlay = $enum5.current;
-          if (overlay.get_name().toLowerCase() !== 'caption' || globalScriptInterface.get_showCaptions()) {
-            overlay.set_tweenFactor(CameraParameters.easeCurve(this._tour.get_currentTourStop().get_tweenPosition(), (overlay.get_interpolationType() === 5) ? this._tour.get_currentTourStop().get_interpolationType() : overlay.get_interpolationType()));
-            overlay.draw3D(renderContext, false);
-          }
-        }
-      }
-      else {
-        var i = 0;
-      }
-      renderContext.restore();
-    }
-  },
-
-  get_tour: function () {
-    return this._tour;
-  },
-
-  set_tour: function (value) {
-    this._tour = value;
-    return value;
-  },
-
-  nextSlide: function () {
-    if (this._tour.get_currentTourStop() != null) {
-      if (!this._tour.get_currentTourStop().get_masterSlide()) {
-        if (this._tour.get_currentTourStop().get_musicTrack() != null) {
-          this._tour.get_currentTourStop().get_musicTrack().stop();
-        }
-        if (this._tour.get_currentTourStop().get_voiceTrack() != null) {
-          this._tour.get_currentTourStop().get_voiceTrack().stop();
-        }
-        var $enum1 = ss.enumerate(this._tour.get_currentTourStop().get_overlays());
-        while ($enum1.moveNext()) {
-          var overlay = $enum1.current;
-          overlay.stop();
-        }
-      }
-      else {
-        this._currentMasterSlide = this._tour.get_currentTourStop();
-      }
-    }
-    if (this._tour.get_currentTourstopIndex() < (this._tour.get_tourStops().length - 1) || this._tour.get_currentTourStop().get_isLinked()) {
-      if (this._tour.get_currentTourStop().get_endTarget() != null) {
-        globalWWTControl.gotoTargetFull(false, true, this._tour.get_currentTourStop().get_endTarget().get_camParams(), this._tour.get_currentTourStop().get_target().get_studyImageset(), this._tour.get_currentTourStop().get_target().get_backgroundImageset());
-        globalWWTControl.set__mover(null);
-      }
-      this._onTarget = false;
-      if (this._tour.get_currentTourStop().get_isLinked()) {
-        try {
-          switch (this._tour.get_currentTourStop().get_nextSlide()) {
-            case 'Return':
-              if (this._callStack.count > 0) {
-                this.playFromTourstop(this._tour.get_tourStops()[this._callStack.pop()]);
-              }
-              else {
-                this._tour.set_currentTourstopIndex(this._tour.get_tourStops().length - 1);
-              }
-              break;
-            default:
-              this.playFromTourstop(this._tour.get_tourStops()[this._tour.getTourStopIndexByID(this._tour.get_currentTourStop().get_nextSlide())]);
-              break;
-          }
-        }
-        catch ($e2) {
-          if (this._tour.get_currentTourstopIndex() < (this._tour.get_tourStops().length - 1)) {
-            this._tour.set_currentTourstopIndex(this._tour.get_currentTourstopIndex() + 1) - 1;
-          }
-        }
-      }
-      else {
-        this._tour.set_currentTourstopIndex(this._tour.get_currentTourstopIndex() + 1) - 1;
-      }
-      if (this._currentMasterSlide != null && this._tour.get_currentTourStop().get_masterSlide()) {
-        this._stopCurrentMaster();
-      }
-      var instant = false;
-      switch (this._tour.get_currentTourStop().get__transition()) {
-        case 0:
-          break;
-        case 1:
-          instant = true;
-          break;
-        case 2:
-          instant = true;
-          break;
-        case 3:
-          instant = true;
-          break;
-        case 5:
-          instant = true;
-          break;
-        case 4:
-          instant = true;
-          break;
-        default:
-          break;
-      }
-      globalWWTControl.gotoTarget(this._tour.get_currentTourStop().get_target(), false, instant, false);
-      this._slideStartTime = ss.now();
-      Settings.tourSettings = this._tour.get_currentTourStop();
-      SpaceTimeController.set_now(this._tour.get_currentTourStop().get_startTime());
-      SpaceTimeController.set_syncToClock(false);
-    } else {
-      this._stopCurrentMaster();
-      TourPlayer._playing = false;
-      if (Settings.get_current().autoRepeatTour) {
-        this._tour.set_currentTourstopIndex(-1);
-        this.play();
-      }
-      else {
-        globalWWTControl._freezeView();
-        if (TourPlayer.__tourEnded != null) {
-          TourPlayer.__tourEnded(this, new ss.EventArgs());
-        }
-        globalWWTControl._hideUI(false);
-        globalScriptInterface._fireTourEnded();
-      }
-    }
-  },
-
-  _stopCurrentMaster: function () {
-    if (this._currentMasterSlide != null) {
-      if (this._currentMasterSlide.get_musicTrack() != null) {
-        this._currentMasterSlide.get_musicTrack().stop();
-      }
-      if (this._currentMasterSlide.get_voiceTrack() != null) {
-        this._currentMasterSlide.get_voiceTrack().stop();
-      }
-      var $enum1 = ss.enumerate(this._currentMasterSlide.get_overlays());
-      while ($enum1.moveNext()) {
-        var overlay = $enum1.current;
-        overlay.stop();
-      }
-      this._currentMasterSlide = null;
-    }
-  },
-
-  get_leaveSettingsWhenStopped: function () {
-    return this._leaveSettingsWhenStopped;
-  },
-
-  set_leaveSettingsWhenStopped: function (value) {
-    this._leaveSettingsWhenStopped = value;
-    return value;
-  },
-
-  play: function () {
-    if (this._tour == null) {
-      return;
-    }
-    if (TourPlayer._playing) {
-      this.stop(true);
-    } else {
-      TourPlayer._playing = true;
-    }
-    globalWWTControl._hideUI(true);
-    TourPlayer._playing = true;
-    if (this._tour.get_tourStops().length > 0) {
-      this._onTarget = false;
-      if (this._tour.get_currentTourstopIndex() === -1) {
-        this._tour.set_currentTourStop(this._tour.get_tourStops()[0]);
-      }
-      var $enum1 = ss.enumerate(this._tour.get_tourStops());
-      while ($enum1.moveNext()) {
-        var stop = $enum1.current;
-        if (stop.get_musicTrack() != null) {
-          stop.get_musicTrack().prepMultimedia();
-        }
-        if (stop.get_voiceTrack() != null) {
-          stop.get_voiceTrack().prepMultimedia();
-        }
-        var $enum2 = ss.enumerate(stop.get_overlays());
-        while ($enum2.moveNext()) {
-          var overlay = $enum2.current;
-          overlay.prepMultimedia();
-        }
-      }
-      if (this._tour.get_currentTourstopIndex() > 0) {
-        this._playMasterForCurrent();
-      }
-      globalWWTControl.gotoTarget(this._tour.get_currentTourStop().get_target(), false, true, false);
-    }
-    this._slideStartTime = ss.now();
-    TourPlayer._playing = true;
-  },
-
-  _playMasterForCurrent: function () {
-    if (!this._tour.get_currentTourStop().get_masterSlide()) {
-      var currentMaster = this._tour.elapsedTimeSinceLastMaster(this._tour.get_currentTourstopIndex());
-      if (currentMaster != null) {
-        var elapsed = currentMaster.duration;
-        this._currentMasterSlide = currentMaster.master;
-        if (this._currentMasterSlide.get_musicTrack() != null) {
-          this._currentMasterSlide.get_musicTrack().seek(elapsed);
-          this._currentMasterSlide.get_musicTrack().play();
-        }
-        if (this._currentMasterSlide.get_voiceTrack() != null) {
-          this._currentMasterSlide.get_voiceTrack().seek(elapsed);
-          this._currentMasterSlide.get_voiceTrack().play();
-        }
-        var $enum1 = ss.enumerate(this._currentMasterSlide.get_overlays());
-        while ($enum1.moveNext()) {
-          var overlay = $enum1.current;
-          overlay.seek(elapsed);
-          overlay.play();
-        }
-      }
-    }
-  },
-
-  stop: function (noSwitchBackFullScreen) {
-    if (TourPlayer._switchedToFullScreen && !noSwitchBackFullScreen) {
-    }
-    if (!this._leaveSettingsWhenStopped) {
-      Settings.tourSettings = null;
-    }
-    TourPlayer._playing = false;
-    if (this._tour.get_currentTourStop() != null) {
-      if (this._tour.get_currentTourStop().get_musicTrack() != null) {
-        this._tour.get_currentTourStop().get_musicTrack().stop();
-      }
-      if (this._tour.get_currentTourStop().get_voiceTrack() != null) {
-        this._tour.get_currentTourStop().get_voiceTrack().stop();
-      }
-      var $enum1 = ss.enumerate(this._tour.get_currentTourStop().get_overlays());
-      while ($enum1.moveNext()) {
-        var overlay = $enum1.current;
-        overlay.stop();
-      }
-    }
-    if (this._currentMasterSlide != null) {
-      if (this._currentMasterSlide.get_musicTrack() != null) {
-        this._currentMasterSlide.get_musicTrack().stop();
-      }
-      if (this._currentMasterSlide.get_voiceTrack() != null) {
-        this._currentMasterSlide.get_voiceTrack().stop();
-      }
-      var $enum2 = ss.enumerate(this._currentMasterSlide.get_overlays());
-      while ($enum2.moveNext()) {
-        var overlay = $enum2.current;
-        overlay.stop();
-      }
-    }
-    globalWWTControl._hideUI(TourPlayer.noRestoreUIOnStop);
-    globalScriptInterface._fireTourEnded();
-  },
-
-  updateSlideStates: function () {
-    var slideChanging = false;
-    var slideElapsedTime = ss.now() - this._slideStartTime;
-    if (slideElapsedTime > this._tour.get_currentTourStop().get_duration() && TourPlayer._playing) {
-      this.nextSlide();
-      slideChanging = true;
-    }
-    slideElapsedTime = ss.now() - this._slideStartTime;
-    if (this._tour.get_currentTourStop() != null) {
-      this._tour.get_currentTourStop().set_tweenPosition(Math.min(1, (slideElapsedTime / this._tour.get_currentTourStop().get_duration())));
-      this._tour.get_currentTourStop().faderOpacity = 0;
-      var elapsedSeconds = this._tour.get_currentTourStop().get_tweenPosition() * this._tour.get_currentTourStop().get_duration() / 1000;
-      if (slideChanging) {
-        globalWWTControl.set_crossFadeFrame(false);
-      }
-      switch (this._tour.get_currentTourStop().get__transition()) {
-        case 0:
-          this._tour.get_currentTourStop().faderOpacity = 0;
-          globalWWTControl.set_crossFadeFrame(false);
-          break;
-        case 2:
-          if (slideChanging) {
-          }
-          if (elapsedSeconds < (elapsedSeconds - this._tour.get_currentTourStop().get__transitionHoldTime())) {
-            globalWWTControl.set_crossFadeFrame(true);
-            this._tour.get_currentTourStop().faderOpacity = 1;
-          }
-          else {
-            this._tour.get_currentTourStop().faderOpacity = 0;
-            globalWWTControl.set_crossFadeFrame(false);
-          }
-          break;
-        case 1:
-          globalWWTControl.set_crossFadeFrame(true);
-          var opacity = Math.max(0, 1 - Math.min(1, (elapsedSeconds - this._tour.get_currentTourStop().get__transitionHoldTime()) / this._tour.get_currentTourStop().get__transitionTime()));
-          this._tour.get_currentTourStop().faderOpacity = opacity;
-          if (slideChanging) {
-          }
-          break;
-        case 3:
-        case 4:
-          globalWWTControl.set_crossFadeFrame(false);
-          var opacity = Math.max(0, 1 - Math.max(0, elapsedSeconds - this._tour.get_currentTourStop().get__transitionHoldTime()) / this._tour.get_currentTourStop().get__transitionTime());
-          this._tour.get_currentTourStop().faderOpacity = opacity;
-          break;
-        case 5:
-          globalWWTControl.set_crossFadeFrame(false);
-          break;
-        default:
-          break;
-      }
-      if (!this._tour.get_currentTourStop().get_isLinked() && this._tour.get_currentTourstopIndex() < (this._tour.get_tourStops().length - 1)) {
-        var nextTrans = this._tour.get_tourStops()[this._tour.get_currentTourstopIndex() + 1].get__transition();
-        var nextTransTime = this._tour.get_tourStops()[this._tour.get_currentTourstopIndex() + 1].get__transitionOutTime();
-        switch (nextTrans) {
-          case 5:
-          case 3:
-            if (!this._tour.get_currentTourStop().faderOpacity) {
-              globalWWTControl.set_crossFadeFrame(false);
-              var opacity = Math.max(0, 1 - Math.min(1, ((this._tour.get_currentTourStop().get_duration() / 1000) - elapsedSeconds) / nextTransTime));
-              this._tour.get_currentTourStop().faderOpacity = opacity;
-            }
-            break;
-          default:
-            break;
-        }
-      }
-    }
-  },
-
-  updateTweenPosition: function (tween) {
-    var slideElapsedTime = ss.now() - this._slideStartTime;
-    if (tween > -1) {
-      return this._tour.get_currentTourStop().set_tweenPosition(Math.min(1, tween));
-    } else {
-      return this._tour.get_currentTourStop().set_tweenPosition(Math.min(1, (slideElapsedTime / this._tour.get_currentTourStop().get_duration())));
-    }
-  },
-
-  close: function () {
-    if (this._tour != null) {
-      if (TourPlayer.get_playing()) {
-        this.stop(TourPlayer._switchedToFullScreen);
-      }
-      this._tour = null;
-    }
-  },
-
-  mouseDown: function (sender, e) {
-    var location;
-    location = this.pointToView(Vector2d.create(e.offsetX, e.offsetY));
-    if (this._tour == null || this._tour.get_currentTourStop() == null) {
-      return false;
-    }
-    for (var i = this._tour.get_currentTourStop().get_overlays().length - 1; i >= 0; i--) {
-      if (this._tour.get_currentTourStop().get_overlays()[i].hitTest(location)) {
-        if (!ss.emptyString(this._tour.get_currentTourStop().get_overlays()[i].get_url())) {
-          var linkItem = this._tour.get_currentTourStop().get_overlays()[i];
-          Util._openUrl(linkItem.get_url());
-          return true;
-        }
-        if (!ss.emptyString(this._tour.get_currentTourStop().get_overlays()[i].get_linkID())) {
-          this._callStack.push(this._tour.get_currentTourstopIndex());
-          this.playFromTourstop(this._tour.get_tourStops()[this._tour.getTourStopIndexByID(this._tour.get_currentTourStop().get_overlays()[i].get_linkID())]);
-          return true;
-        }
-      }
-    }
-    return false;
-  },
-
-  mouseUp: function (sender, e) {
-    return false;
-  },
-
-  mouseMove: function (sender, e) {
-    var location;
-    try {
-      location = this.pointToView(Vector2d.create(e.offsetX, e.offsetY));
-    }
-    catch ($e1) {
-      return false;
-    }
-    if (this._tour == null || this._tour.get_currentTourStop() == null) {
-      return false;
-    }
-    for (var i = this._tour.get_currentTourStop().get_overlays().length - 1; i >= 0; i--) {
-      if (this._tour.get_currentTourStop().get_overlays()[i].hitTest(location) && (!ss.emptyString(this._tour.get_currentTourStop().get_overlays()[i].get_url()) || !ss.emptyString(this._tour.get_currentTourStop().get_overlays()[i].get_linkID()))) {
-        return true;
-      }
-    }
-    return false;
-  },
-
-  mouseClick: function (sender, e) {
-    return false;
-  },
-
-  click: function (sender, e) {
-    return false;
-  },
-
-  mouseDoubleClick: function (sender, e) {
-    return false;
-  },
-
-  keyDown: function (sender, e) {
-    switch (e.keyCode) {
-      case 27:
-        this.stop(TourPlayer._switchedToFullScreen);
-        globalWWTControl._closeTour();
-        return true;
-      case 32:
-        this.pauseTour();
-        return true;
-      case 39:
-        this._playNextSlide();
-        return true;
-      case 37:
-        this._playPreviousSlide();
-        return true;
-      case 35:
-        if (this._tour.get_tourStops().length > 0) {
-          this.playFromTourstop(this._tour.get_tourStops()[this._tour.get_tourStops().length - 1]);
-        }
-        return true;
-      case 36:
-        if (this._tour.get_tourStops().length > 0) {
-          this.playFromTourstop(this._tour.get_tourStops()[0]);
-        }
-        return true;
-    }
-    return false;
-  },
-
-  _playNextSlide: function () {
-    if ((this._tour.get_currentTourstopIndex() < this._tour.get_tourStops().length - 1) && this._tour.get_tourStops().length > 0) {
-      this.playFromTourstop(this._tour.get_tourStops()[this._tour.get_currentTourstopIndex() + 1]);
-    }
-  },
-
-  _playPreviousSlide: function () {
-    if (this._tour.get_currentTourstopIndex() > 0) {
-      this.playFromTourstop(this._tour.get_tourStops()[this._tour.get_currentTourstopIndex() - 1]);
-    }
-  },
-
-  playFromTourstop: function (tourStop) {
-    this.stop(true);
-    this._tour.set_currentTourStop(tourStop);
-    globalWWTControl.gotoTarget(this._tour.get_currentTourStop().get_target(), false, true, false);
-    SpaceTimeController.set_now(this._tour.get_currentTourStop().get_startTime());
-    SpaceTimeController.set_syncToClock(false);
-    this.play();
-  },
-
-  pauseTour: function () {
-    if (TourPlayer._playing) {
-      this.stop(TourPlayer._switchedToFullScreen);
-      globalWWTControl._freezeView();
-      globalScriptInterface._fireTourPaused();
-    } else {
-      this.play();
-      globalScriptInterface._fireTourResume();
-    }
-  },
-
-  keyUp: function (sender, e) {
-    return false;
-  },
-
-  hover: function (pnt) {
-    if (TourPlayer._playing) {
-      return true;
-    }
-    return false;
-  },
-
-  pointToView: function (pnt) {
-    var clientHeight = globalWWTControl.canvas.height;
-    var clientWidth = globalWWTControl.canvas.width;
-    var viewWidth = (clientWidth / clientHeight) * 1116;
-    var x = ((pnt.x) / (clientWidth) * viewWidth) - ((viewWidth - 1920) / 2);
-    var y = (pnt.y) / clientHeight * 1116;
-    return Vector2d.create(x, y);
-  }
-};
-
-registerType("TourPlayer", [TourPlayer, TourPlayer$, null, IUiController]);
-
-// wwtlib.MasterTime
-
-export function MasterTime(master, duration) {
-  this.duration = 0;
-  this.master = master;
-  this.duration = duration;
-}
-
-var MasterTime$ = {};
-
-registerType("MasterTime", [MasterTime, MasterTime$, null]);
 
 // wwtlib.FolderDownloadAction
